experiments/requirements.txt
# Core dependencies for PM Agent experiments
datasets>=2.14.0
huggingface-hub>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# SWE-bench specific
swebench>=1.0.0
transformers>=4.35.0

# PM Agent dependencies
asyncio>=3.4.3
aiohttp>=3.9.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.0
py-cpuinfo>=9.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# Git operations
gitpython>=3.1.40

# Analysis and reporting
jupyter>=1.0.0
plotly>=5.18.0
tabulate>=0.9.0
jinja2>=3.1.0

# Testing infrastructure
faker>=20.0.0
factory-boy>=3.3.0
hypothesis>=6.90.0

# API clients (for cost tracking)
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
structlog>=23.2.0